+++ teachertool/src/components/DebugInput.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Button } from "react-common/components/controls/Button";
 import { Input } from "react-common/components/controls/Input";
 import { Textarea } from "react-common/components/controls/Textarea";
-import { runEvaluateAsync } from "../transforms/runEvaluate";
+import { runEvaluateAsync } from "../transforms/runEvaluateAsync";
 
 interface IProps {}
